+++ TSPInstance.hpp
@@ -2,7 +2,9 @@
 #define TSPINSTANCE_H
 
 #include <vector>
+
 #include "types.hpp"
+#include "Graph.hpp"
 
 class TSPInstance {
 private:
@@ -11,14 +13,14 @@ private:
     EdgeWeightType type;
 
     std::vector<Point> points;
-    std::vector<std::vector<double>> adj_matrix;
+    std::vector<std::vector<int>> adj_matrix;
 
 public:
     TSPInstance(
         int _dimension, 
         EdgeWeightType _type, 
         std::vector<Point> _points, 
-        std::vector<std::vector<double>> _adj_matrix
+        std::vector<std::vector<int>> _adj_matrix
     );
 };
